requirements.txt
apache-airflow-providers-apache-spark>=2.0.0
pyspark>=3.4.0
apache-airflow-providers-openlineage>=1.8.0
apache-airflow==2.7.1
pandas
sqlalchemy
apache-airflow-providers-amazon
apache-airflow-providers-snowflake
boto3
python-dotenv
selenium
webdriver-manager